requirements.txt
# Google Sheets API
gspread==6.2.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.2.0

# HTTP requests for Telegram API
requests==2.31.0

# Environment variables (for local development)
python-dotenv==1.0.0

# AI Integration Dependencies
openai>=1.0.0
google-generativeai>=0.3.0
